requirements.txt
ansible
# behave
boto3
configparser
pywinrm
testinfra